header, footer, and show accuracy for FA-Cup and Premier League
